% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqn_beta.R
\name{eqn_beta}
\alias{eqn_beta}
\title{"eqn_beta" estimate $beta$ for the lagrange multipliers}
\usage{
eqn_beta(x, N_0, S_0)
}
\arguments{
\item{x}{}

\item{N_0}{}

\item{S_0}{}
}
\description{
"eqn_beta" estimate $beta$ for the lagrange multipliers
}
