% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negLL_functions.R
\name{negLL_com_pmete}
\alias{negLL_com_pmete}
\title{negLL_com_pmete}
\usage{
negLL_com_pmete(par, x, xmin, beta, N_0, S_0)
}
\arguments{
\item{par}{}

\item{x}{}

\item{xmin}{}

\item{beta}{}

\item{N_0}{}

\item{S_0}{}
}
\description{
Deriving the log-likelihood function for the individual size distribution to solve \eqn{\beta_{c}} and \eqn{\lambda_1} and \eqn{\lambda_2}
}
